Add unit tests for StarSchemaDetector and Vocabulary functionality
- Implement tests for StarSchemaDetector to validate hub/spoke detection logic, including scenarios for multiple spokes, unmatched tables, and independent hubs.
- Introduce VocabularyTests to cover set operations, containment, and vocabulary accumulation for various data types.
- Update CsvOutputWriterTests to ensure correct data type handling for CSV output.

+++ tests/DatumIngest.Tests/Catalog/GzipProviderTests.cs
@@ -80,7 +80,7 @@ public sealed class GzipProviderTests : IDisposable
         Assert.Equal("age", schema.Columns[1].Name);
         Assert.Equal("score", schema.Columns[2].Name);
         Assert.Equal(DataKind.String, schema.Columns[0].Kind);
-        Assert.Equal(DataKind.Int8, schema.Columns[1].Kind);
+        Assert.Equal(DataKind.Int32, schema.Columns[1].Kind);
         Assert.Equal(DataKind.Float64, schema.Columns[2].Kind);
     }
